Pipeline manager updated
