+++ CrowEngine/include/Utils/FileUtilities.hpp
@@ -6,7 +6,6 @@
 #include <unordered_set>
 #include <fstream>
 
-
 template<typename T>
 void RemoveDuplicatesInList(std::vector<T> &list)
 {
@@ -22,16 +21,4 @@ void RemoveDuplicatesInList(std::vector<T> &list)
   list.resize(index);  // Resize the list to remove extra elements
 }
 
-std::vector<char> ReadFile(const std::string &filepath)
-{
-  std::ifstream file{filepath, std::ios::ate | std::ios::binary};
-  if (!file.is_open()) {
-    throw std::runtime_error("failed to open file: " + filepath);
-  }
-  size_t fileSize = static_cast<size_t>(file.tellg());
-  std::vector<char> buffer(fileSize);
-  file.seekg(0);
-  file.read(buffer.data(), fileSize);
-  file.close();
-  return buffer;
-}
+std::vector<char> ReadFile(const std::string &filepath);
