+++ CrowEngine/include/crow_utils.hpp
@@ -6,5 +6,4 @@ namespace crowe
 {
   void RemoveDuplicatesInList(std::vector<uint32_t> &intList);
 
-  std::vector<char> ReadFile(const std::string &filepath);
 }
