+++ CrowEngine/src/Utils/FileUtilities.cpp
@@ -1,0 +1,15 @@
+#include "Utils/FileUtilities.hpp"
+
+std::vector<char> ReadFile(const std::string &filepath)
+{
+  std::ifstream file{filepath, std::ios::ate | std::ios::binary};
+  if (!file.is_open()) {
+    throw std::runtime_error("failed to open file: " + filepath);
+  }
+  size_t fileSize = static_cast<size_t>(file.tellg());
+  std::vector<char> buffer(fileSize);
+  file.seekg(0);
+  file.read(buffer.data(), fileSize);
+  file.close();
+  return buffer;
+}
